chore: deep linking fixes, refactor factories
